F1-Backend-Django/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
